Creating new class for FakeInputSplit to avoid java doc warnings due to JDK-8176901
